Tests/LibWeb/Text/expected/css/typed-om-set-invalid-value-preserves-existing.txt

Initial style: width: 10px;
Invalid set threw TypeError
After invalid set: width: 10px;
